feat: don't copy symlinks into old

+++ makesymlinks.sh
@@ -6,11 +6,13 @@
 
 ########## Variables
 
+dotfiles_dir=~/Projects/dotfiles
+
 dir=~/Projects/dotfiles                    # dotfiles directory
 olddir=~/Projects/dotfiles_old             # old dotfiles backup directory
 files="bashrc zshrc oh-my-zsh gitconfig"    # list of files/folders to symlink in homedir
 
-
+bin_dir=$dotfiles_dir/bin
 
 # create dotfiles_old in homedir
 echo -n "Creating $olddir for backup of any existing dotfiles in ~ ..."
@@ -24,11 +26,27 @@ echo "done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file $olddir
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [ -e ~/.$file ]; then
+	    if [ ! -L ~/.$file ]; then
+	        echo "Moving any existing dotfiles from ~ to $olddir"
+    	    mv ~/.$file $olddir
+            echo "Creating symlink to $file in home directory."
+            ln -s $dir/$file ~/.$file
+	    else 
+	        echo "File ~/.$file is a symlink" 
+	    fi
+    else
+    	echo "File ~/.$file doesn't exist"
+	    echo "Creating symlink to $file in home directory."
+     	ln -s $dir/$file ~/.$file
+    fi
 done
+
+for file in ${bin_dir}/*; do
+    [ -e "$file" ] || continue
+    echo "Bin dir file $file"
+done
+
 
 install_zsh () {
 # Test to see if zshell is installed.  If it is:
